タイムラインのコンポーネント化，特定ユーザの投稿のみ取得可能に
